move get_compute_cost function to stage_profiling
